fix argument bug

+++ main.py
@@ -60,7 +60,7 @@ if __name__ == "__main__":
     parser.add_argument("--value_support_delta", default=0.01, type=float)
     parser.add_argument("--m_top", default=8, type=int)
     parser.add_argument("--c_visit", default=36, type=int)
-    parser.add_argument("--c_scale", default=0.1, type=int)
+    parser.add_argument("--c_scale", default=0.1, type=float)
     args = parser.parse_args()
 
     set_seed(args.seed)
